# public/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Board API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "host": "lara.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/adverts": {
            "get": {
                "tags": [
                    "Adverts"
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AdvertList"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": [],
                        "OAuth2": []
                    }
                ]
            }
        },
        "/adverts/{advertId}": {
            "get": {
                "tags": [
                    "Adverts"
                ],
                "parameters": [
                    {
                        "name": "advertId",
                        "in": "path",
                        "description": "ID of advert",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/AdvertDetail"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": [],
                        "OAuth2": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "AdvertDetail": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "user": {
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "category": {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "region": {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "date": {
                    "properties": {
                        "published": {
                            "type": "date"
                        },
                        "expires": {
                            "type": "date"
                        }
                    },
                    "type": "object"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AdvertValue"
                    }
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            },
            "type": "object"
        },
        "AdvertValue": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            },
            "type": "object"
        },
        "AdvertList": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "user": {
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "category": {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "region": {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                },
                "title": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "date": {
                    "type": "date"
                },
                "photo": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "OAuth2": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://lara.com/oauth/token"
        },
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}